// src/app/Model/user.ts
export interface User {

   password: string;
   lastLogin: Date;
   emailId: string;
   contactNumber: string;
   _id: string;
   role: string;
   tallyUserName: string;
   posCashLedgerName: string;
   cashLedgerName: string;
   basicBasePartyName: string;
   placeOfSupply: string;
   voucherTypeName: string;
   godownName: string;
   salesVoucherHeader: string;
   salesVoucherFooter: string;

}